% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/security_info.R
\name{security_info}
\alias{security_info}
\title{Access the Treasury Direct Security endpoint}
\usage{
security_info(cusip = NULL, issue_date = NULL)
}
\arguments{
\item{cusip}{security cusip, must be 9 characters long}

\item{issue_date}{date security was issued, must be in "YYYY-MM--DD" format, character input will be coerced to date}
}
\value{
the results from the searcha tibble of treasury security information
}
\description{
Returns a large collection of variables relating to
a specific treasury security including bid to cover ratio, rates,
auction data, and other reference information
}
\examples{

# get data for 2-year Treasury
security_info("91282CBX8", "2021-04-30")
}
